Reorganize project, added safeguard for year, documentation, and remove NuGet
